+++ pennylane/ops/__init__.py
@@ -39,7 +39,6 @@ from .channel import __all__ as _channel__ops__
 
 class AdjointError(Exception):
     """Exception for non-adjointable operations."""
-
     pass