+++ .github/CHANGELOG.md
@@ -285,7 +285,7 @@
 
 <h3>Improvements</h3>
 
-* The `./pennylane/ops/qubit.py` file is broken up into a folder of eight separate files.
+* The `./pennylane/ops/qubit.py` file is broken up into a folder of six separate files.
   [(#1467)](https://github.com/PennyLaneAI/pennylane/pull/1467)
 
 * Changed to using commas as the separator of wires in the string